# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: VN Fashion Shop API
  description: |
    RESTful API cho nen tang thuong mai dien tu thoi trang Viet Nam.
    RESTful API for the Vietnamese fashion e-commerce platform.

    ## Xac thuc / Authentication
    Su dung Bearer JWT token trong header `Authorization`.
    Use Bearer JWT tokens in the `Authorization` header.

    ## Phan trang / Pagination
    Cac endpoint tra ve danh sach su dung query params `page` va `page_size`.
    List endpoints use `page` and `page_size` query parameters.

    ## Ma loi / Error Codes
    - `400` -- Bad Request
    - `401` -- Unauthorized (chua xac thuc)
    - `403` -- Forbidden (khong co quyen)
    - `404` -- Not Found
    - `422` -- Validation Error
    - `500` -- Internal Server Error
  version: "1.0.0"
  contact:
    name: VN Fashion Shop Team
    email: dev@vnfashion.vn
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.vnfashion.vn
    description: Production
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token nhan duoc tu endpoint /auth/login
  schemas:
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: "Khong tim thay san pham"
    ValidationError:
      type: object
      required: [detail]
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string

    # ---- Auth ----
...
    CategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Ao"
        slug:
          type: string
          example: "ao"
        description:
          type: string
          example: "Cac loai ao thoi trang nam nu"
        image_url:
          type: string
          nullable: true
        parent_id:
          type: string
          format: uuid
          nullable: true
        sort_order:
          type: integer
        is_active:
          type: boolean
        children:
          type: array
          items:
            $ref: "#/components/schemas/CategoryResponse"
...
    ProductSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Ao Dai Hoa Sen"
        slug:
          type: string
          example: "ao-dai-hoa-sen"
        short_description:
          type: string
          example: "Ao dai truyen thong voi hoa van hoa sen"
        base_price:
          type: number
          format: double
          example: 1500000
        sale_price:
          type: number
          format: double
          nullable: true
          example: 1200000
        currency:
          type: string
          example: "VND"
        primary_image_url:
          type: string
        is_featured:
          type: boolean
        average_rating:
          type: number
          format: double
          example: 4.5
        review_count:
          type: integer
          example: 23
    ProductDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Ao Dai Hoa Sen"
        slug:
          type: string
          example: "ao-dai-hoa-sen"
        description:
          type: string
          example: "Ao dai truyen thong Viet Nam duoc thiet ke voi hoa van hoa sen tinh te, chat lieu lua to tam cao cap."
        short_description:
          type: string
          example: "Ao dai truyen thong voi hoa van hoa sen"
        base_price:
          type: number
          format: double
          example: 1500000
        sale_price:
          type: number
          format: double
          nullable: true
          example: 1200000
        currency:
          type: string
          example: "VND"
        brand:
          type: string
          example: "Thuy Design House"
        material:
          type: string
          example: "Lua to tam 100%"
        care_instructions:
          type: string
          example: "Giat tay nhe, khong vat, phoi bong mat"
        origin:
          type: string
          example: "Viet Nam"
        is_active:
          type: boolean
        is_featured:
          type: boolean
        total_sold:
          type: integer
          example: 150
        average_rating:
          type: number
          format: double
          example: 4.5
        review_count:
          type: integer
          example: 23
        category:
          $ref: "#/components/schemas/CategoryResponse"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          example: "https://cdn.vnfashion.vn/products/ao-dai-hoa-sen-1.jpg"
        alt_text:
          type: string
          example: "Ao dai hoa sen - mat truoc"
        sort_order:
          type: integer
        is_primary:
          type: boolean
    ProductVariant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
          example: "ADHS-DO-M"
        size:
          type: string
          enum: [XS, S, M, L, XL, XXL]
          example: "M"
        color:
          type: string
          example: "Do"
        color_hex:
          type: string
          example: "#DC143C"
        price:
          type: number
          format: double
          example: 1200000
        stock_quantity:
          type: integer
          example: 25
        is_active:
          type: boolean
    CreateProductRequest:
      type: object
      required: [name, category_id, base_price]
      properties:
        name:
          type: string
          example: "Ao Dai Hoa Sen"
        category_id:
          type: string
          format: uuid
        description:
          type: string
          example: "Ao dai truyen thong Viet Nam duoc thiet ke voi hoa van hoa sen."
        short_description:
          type: string
        base_price:
          type: number
          format: double
          example: 1500000
        sale_price:
          type: number
          format: double
          nullable: true
        brand:
          type: string
        material:
          type: string
        care_instructions:
          type: string
        origin:
          type: string
          default: "Viet Nam"
        is_featured:
          type: boolean
          default: false
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        category_id:
          type: string
          format: uuid
        description:
          type: string
        short_description:
          type: string
        base_price:
          type: number
          format: double
        sale_price:
          type: number
          format: double
          nullable: true
        brand:
          type: string
        material:
          type: string
        is_active:
          type: boolean
        is_featured:
          type: boolean
    PaginatedProducts:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProductSummary"
        total:
          type: integer
          example: 120
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 6

    # ---- Cart ----
...
paths:
...
  /categories:
    get:
      tags: [Categories]
      summary: Danh sach danh muc / List categories
      security: []
      operationId: listCategories
      parameters:
        - name: include_inactive
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Danh sach danh muc (dang cay)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponse"
  /categories/{slug}:
    get:
      tags: [Categories]
      summary: Chi tiet danh muc / Category detail
      security: []
      operationId: getCategory
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          example: "ao-dai"
      responses:
        "200":
          description: Thong tin danh muc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "404":
          description: Khong tim thay danh muc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ======== Products ========
  /products:
    get:
      tags: [Products]
      summary: Danh sach san pham / List products
      security: []
      operationId: listProducts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: category
          in: query
          description: Loc theo slug danh muc
          schema:
            type: string
          example: "ao-dai"
        - name: search
          in: query
          description: Tim kiem theo ten san pham
          schema:
            type: string
          example: "ao dai"
        - name: min_price
          in: query
          schema:
            type: number
            format: double
          example: 500000
        - name: max_price
          in: query
          schema:
            type: number
            format: double
          example: 2000000
        - name: size
          in: query
          schema:
            type: string
            enum: [XS, S, M, L, XL, XXL]
        - name: color
          in: query
          schema:
            type: string
          example: "Do"
        - name: sort
          in: query
          description: Sap xep / Sort order
          schema:
            type: string
            enum: [newest, price_asc, price_desc, best_selling, top_rated]
            default: newest
        - name: is_featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Danh sach san pham phan trang
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProducts"
  /products/{slug}:
    get:
      tags: [Products]
      summary: Chi tiet san pham / Product detail
      security: []
      operationId: getProduct
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          example: "ao-dai-hoa-sen"
      responses:
        "200":
          description: Thong tin chi tiet san pham
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Khong tim thay san pham
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: "Khong tim thay san pham"

  # ======== Cart ========
...
  /admin/products:
    post:
      tags: [Admin]
      summary: Them san pham moi / Create product
      operationId: adminCreateProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Tao san pham thanh cong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "401":
          description: Chua xac thuc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Khong co quyen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Du lieu khong hop le
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
  /admin/products/{product_id}:
    patch:
      tags: [Admin]
      summary: Cap nhat san pham / Update product
      operationId: adminUpdateProduct
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Cap nhat thanh cong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "401":
          description: Chua xac thuc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Khong co quyen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Khong tim thay san pham
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Admin]
      summary: Xoa san pham / Delete product
      operationId: adminDeleteProduct
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Xoa thanh cong
        "401":
          description: Chua xac thuc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Khong co quyen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Khong tim thay san pham
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...